Introduce DeactivationTestingGrain
Introduce DeactivationTestingGrain for reliable, awaitable grain deactivation in tests, replacing Thread.Sleep with WaitForDeactivationAsync. Move all deactivation tracing tests to new DeactivationTracingTests.cs, updating test grains to inherit from DeactivationTestingGrain. Split trace context propagation tests into GrainCallTraceContextPropagationTests.cs for better organization. Refactor ActivationTracingTests.cs to focus on activation and context propagation only. Improves test determinism, reliability, and clarity.
